refactor(cmd): rename cmd/dmsg/dmsgpty-{host,ui,cli} → pty-* (final pty rename)
Last piece of the deeper dmsgpty → pty rename. Three cmd
directories renamed; import paths in the three external
references (cmd/dmsg/dmsg/commands/root.go, cmd/apps/pty/
commands/pty.go, the pkg/visor/init_dmsg.go comment) updated in
lockstep. Internal self-imports (each main package importing its
own commands subpackage) follow the dir move automatically.

Top-level main files in each dir also renamed for consistency
(cmd/dmsg/pty-host/dmsgpty-host.go → pty-host.go, etc.). Standalone
binary names follow the dir name on `go build .` — old binaries
no longer named "dmsgpty-host"/etc. Operators using the unified
CLI surface (`skywire app pty {dmsg,tcp,http,exec}` from #2864/
#2866 era) are unaffected; operators invoking the standalone
binaries directly need to update muscle memory to `pty-host` etc.
The `skywire dmsg pty {cli,host,ui}` subtree is unchanged
(still mounted on the dmsg standalone binary; Hidden in skywire).

Verified locally:
- Standalone /tmp/pty-host, pty-ui, pty-cli build + show help
- `skywire app pty --help` shows the four-mode listing
- pkg/pty + pkg/visor + pkg/visor/visorconfig tests green

The unified-app-framework + dmsgpty → pty rename track is
complete with this PR. RFC #2863 covered the high-level
restructure; the rename closes out the code-level naming.
