+++ cmd/dmsg/dmsg/commands/root.go
@@ -17,9 +17,9 @@ import (
 	dc "github.com/skycoin/skywire/cmd/dmsg/dmsgcurl/commands"
 	dh "github.com/skycoin/skywire/cmd/dmsg/dmsghttp/commands"
 	di "github.com/skycoin/skywire/cmd/dmsg/dmsgip/commands"
-	dpc "github.com/skycoin/skywire/cmd/dmsg/dmsgpty-cli/commands"
-	dph "github.com/skycoin/skywire/cmd/dmsg/dmsgpty-host/commands"
-	dpu "github.com/skycoin/skywire/cmd/dmsg/dmsgpty-ui/commands"
+	dpc "github.com/skycoin/skywire/cmd/dmsg/pty-cli/commands"
+	dph "github.com/skycoin/skywire/cmd/dmsg/pty-host/commands"
+	dpu "github.com/skycoin/skywire/cmd/dmsg/pty-ui/commands"
 	dw "github.com/skycoin/skywire/cmd/dmsg/dmsgweb/commands"
 	dsp "github.com/skycoin/skywire/cmd/dmsg/self-ping/commands"
 	"github.com/skycoin/skywire/pkg/buildinfo"
