+++ cmd/dmsg/pty-host/commands/confgen.go
